backend/pyproject.toml
[project]
name = "test-generator-backend"
version = "1.0.0"
description = "AI Test Generator - FastAPI Backend"
requires-python = ">=3.10,<3.14"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.1",
    "google-generativeai==0.8.3",
    "python-dotenv==1.0.1",
    "pydantic==2.10.3",
]

[tool.uv]
package = false